+++ CHANGELOG.md
@@ -1,15 +1,33 @@
-Change Log
-==========
+# Release Notes
 
-### 07/07/2018 - 1.2.0
+
+## v1.3.0 (07/07/2018)
+
+### Added
+* Added new method `random()` on `Collection`
+* Added new method `only()` and `except()` on `Collection`
+
+## v1.2.0 (07/07/2018)
+
+### Added
 * Added new `Collection` methods, `where()`, `whereIn()` and `filter()`
 * Added protected method on `Collection` for `filterWhere()` on an operator check.
 
-### 07/06/2018 - 1.1.0
-* Added new `Collection` methods, `sort()` and `sortBy()` (ability to sort the items in collection)
-* Added protected methods on `Collection` for `isCallable()` and `ValueCallable`.
 
-### 06/07/2018 - 1.0.1
-* Minor updates to `Arr` class and added `Arr` tests for code better code coverage.
-* Changed the `collection->has()` to work with "dot" notation.
+## v1.1.0 (07/06/2018)
+
+### Added
+* Added new `Collection` methods, `sort()` and `sortBy()` (ability to sort the items in collection)
+* Added protected methods on `Collection` for `isCallable()` and `ValueCallable()`.
+
+
+## v1.0.1 (06/07/2018)
+
+### Added
 * Added more `Collection` tests to improve code coverage.
+
+### Changed
+* Changed the `collection->has()` to work with "dot" notation.
+
+### Fixed
+* Minor updates to `Arr` class and added `Arr` tests for code better code coverage.